requirements.txt
kafka-python==2.0.2
faker==20.1.0
pyspark==3.5.0
